requirements.txt
pandas
pyarrow
gcsfs
google-cloud-storage
google-cloud-bigquery
apache-airflow[gcp]==2.9.1
requests
